Use HTML/JS/CSS functions extracted in zimscraperlib
