+++ pyproject.toml
@@ -12,7 +12,8 @@ readme = "README.md"
 dependencies = [
   "warcio==1.7.4",
   "requests==2.32.3",
-  "zimscraperlib==4.0.0",
+  # use zimscraperlib pinned version once content rewriting functions have been released
+  "zimscraperlib @ git+https://github.com/openzim/python-scraperlib@main",
   "jinja2==3.1.4", # also update version in build-system above and in build_js.sh
   # to support possible brotli content in warcs, must be added separately
   "brotlipy==0.7.0",
@@ -23,6 +24,9 @@ dependencies = [
   "python-dateutil==2.9.0.post0",
 ]
 dynamic = ["authors", "classifiers", "keywords", "license", "version", "urls"]
+
+[tool.hatch.metadata]
+allow-direct-references = true
 
 [tool.hatch.metadata.hooks.openzim-metadata]
 kind = "scraper"
